+++ Hardware/Encoder.h
@@ -2,7 +2,7 @@
 #define __ENCODER_H
 
 void Encoder_Init(void);
-int16_t Encoder1_Get(void);
-int16_t Encoder2_Get(void);
+int32_t Encoder1_Get(void);
+int32_t Encoder2_Get(void);
 
 #endif
